+++ src/main/java/net/lopymine/mtd/mixin/ScreenEffectRendererMixin.java
@@ -2,9 +2,12 @@ package net.lopymine.mtd.mixin;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.*;
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.lopymine.mtd.config.MyTotemDollConfig;
+import net.lopymine.mtd.config.totem.TotemDollSkinType;
 import net.lopymine.mtd.doll.renderer.*;
 import net.lopymine.mtd.utils.LightningUtils;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.renderer.*;
 import net.minecraft.client.renderer.item.ItemStackRenderState;
 import net.minecraft.world.item.ItemStack;
@@ -32,7 +35,8 @@ public class ScreenEffectRendererMixin {
 			return;
 		}
 		LightningUtils.flat();
-		TotemDollRenderer.submitItemAnyway(submitNodeCollector, poseStack, DollRenderContext.D_FLOATING, stack, lightCoords, overlayCoords, outlineColor);
+		AbstractClientPlayer holdingPlayer = MyTotemDollConfig.getInstance().getStandardTotemDollSkinType() == TotemDollSkinType.HOLDING_PLAYER ? Minecraft.getInstance().player : null;
+		TotemDollRenderer.submitItemAnyway(submitNodeCollector, poseStack, DollRenderContext.D_FLOATING, stack, holdingPlayer, lightCoords, overlayCoords, outlineColor);
 	}
 
 }